Fix the build errors related to partitioning
